requirements.txt
# Core Framework
Flask>=2.3.3
Werkzeug>=3.0.1
Jinja2>=3.1.2
click>=8.1.7
itsdangerous>=2.1.2
MarkupSafe>=2.1.3

# Database and Migrations
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
SQLAlchemy>=2.0.23
alembic>=1.9.0
greenlet>=2.0.2

# Authentication and Forms
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
WTForms>=3.2.1
email-validator>=2.1.0.post1

# Environment and Utils
python-dotenv>=1.0.0
blinker>=1.8.2

# Windows Compatibility
colorama>=0.4.6  # Required for Windows colored output
